requirements.txt
pygame>=2.5
Pillow>=10.0
pytmx>=3.31
